web-requirements.txt
flask
flask-login
flask-sqlalchemy
sqlalchemy-migrate
flask-wtf

